Move model to searcher
